Move widget data template to its own control
