+++ tools/Dashboard/DevHome.Dashboard/DevHome.Dashboard.csproj
@@ -13,6 +13,7 @@
     <None Remove="Assets\HostConfigLight.json" />
     <None Remove="Views\AddWidgetDialog.xaml" />
     <None Remove="Views\DashboardView.xaml" />
+    <None Remove="Views\WidgetControl.xaml" />
   </ItemGroup>
 
   <ItemGroup>
@@ -36,6 +37,12 @@
     <Content Update="Assets\HostConfigDark.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Page Update="Views\WidgetControl.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
   </ItemGroup>
 
   <ItemGroup>
