+++ tools/Dashboard/DevHome.Dashboard/Views/DashboardView.xaml
@@ -11,7 +11,6 @@
     xmlns:commonviews="using:DevHome.Common.Views"
     xmlns:vm="using:DevHome.Dashboard.ViewModels"
     xmlns:views="using:DevHome.Dashboard.Views"
-    xmlns:helpers="using:DevHome.Dashboard.Helpers"
     mc:Ignorable="d">
 
     <Grid>
@@ -39,16 +38,7 @@
                 <GridView x:Name="MainGridView" ItemsSource="{x:Bind views:DashboardView.PinnedWidgets}">
                     <GridView.ItemTemplate>
                         <DataTemplate x:DataType="vm:WidgetViewModel">
-                            <Grid Width="300" Height="{x:Bind helpers:WidgetHelpers.GetPixelHeightFromWidgetSize(WidgetSize), Mode=OneWay}">
-                                <ContentPresenter Content="{x:Bind WidgetUIElement, Mode=OneWay}"/>
-                                <Button Tag="{x:Bind}" FontFamily="{StaticResource SymbolThemeFontFamily}" Content="&#xE712;" 
-                                        VerticalAlignment="Top" HorizontalAlignment="Right" Margin="0 30 3 0" BorderThickness="0"
-                                        Click="OpenWidgetMenu">
-                                    <Button.Flyout>
-                                        <MenuFlyout/>
-                                    </Button.Flyout>
-                                </Button>
-                            </Grid>
+                            <views:WidgetControl WidgetSource="{x:Bind}" />
                         </DataTemplate>
                     </GridView.ItemTemplate>
                     <GridView.ItemsPanel>
