+++ tools/Dashboard/DevHome.Dashboard/Views/WidgetControl.xaml
@@ -1,0 +1,24 @@
+<!-- Copyright (c) Microsoft Corporation and Contributors. -->
+<!-- Licensed under the MIT License. -->
+
+<UserControl
+    x:Class="DevHome.Dashboard.Views.WidgetControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:DevHome.Dashboard.Views"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:helpers="using:DevHome.Dashboard.Helpers"
+    mc:Ignorable="d">
+
+    <Grid Width="300" Height="{x:Bind helpers:WidgetHelpers.GetPixelHeightFromWidgetSize(WidgetSource.WidgetSize), Mode=OneWay}">
+        <ContentPresenter Content="{x:Bind WidgetSource.WidgetUIElement, Mode=OneWay}"/>
+        <Button Tag="{x:Bind}" FontFamily="{StaticResource SymbolThemeFontFamily}" Content="&#xE712;" 
+                VerticalAlignment="Top" HorizontalAlignment="Right" Margin="0 30 3 0" BorderThickness="0"
+                Click="OpenWidgetMenu">
+            <Button.Flyout>
+                <MenuFlyout/>
+            </Button.Flyout>
+        </Button>
+    </Grid>
+</UserControl>
